hook most of do exchange up
