+++ grpc-api/src/main/java/io/deephaven/grpc_api/arrow/ArrowFlightUtil.java
@@ -4,23 +4,32 @@ import com.google.common.io.LittleEndianDataInputStream;
 import com.google.protobuf.CodedInputStream;
 import com.google.protobuf.WireFormat;
 import com.google.rpc.Code;
+import dagger.assisted.Assisted;
+import dagger.assisted.AssistedFactory;
+import dagger.assisted.AssistedInject;
 import gnu.trove.iterator.TLongIterator;
 import gnu.trove.list.array.TLongArrayList;
 import io.deephaven.UncheckedDeephavenException;
+import io.deephaven.barrage.flatbuf.BarrageMessageType;
 import io.deephaven.barrage.flatbuf.BarrageMessageWrapper;
+import io.deephaven.barrage.flatbuf.BarrageSubscriptionRequest;
+import io.deephaven.configuration.Configuration;
 import io.deephaven.datastructures.util.CollectionUtil;
 import io.deephaven.db.tables.Table;
 import io.deephaven.db.util.LongSizedDataStructure;
 import io.deephaven.db.util.liveness.SingletonLivenessManager;
+import io.deephaven.db.v2.QueryTable;
 import io.deephaven.db.v2.sources.chunk.ChunkType;
 import io.deephaven.db.v2.utils.BarrageMessage;
 import io.deephaven.db.v2.utils.Index;
 import io.deephaven.db.v2.utils.IndexShiftData;
+import io.deephaven.grpc_api.barrage.BarrageMessageProducer;
 import io.deephaven.grpc_api.barrage.BarrageStreamGenerator;
 import io.deephaven.grpc_api.barrage.util.BarrageSchemaUtil;
 import io.deephaven.grpc_api.session.SessionState;
 import io.deephaven.grpc_api.session.TicketRouter;
 import io.deephaven.grpc_api.util.BrowserStream;
+import io.deephaven.grpc_api.util.ExportTicketHelper;
 import io.deephaven.grpc_api.util.GrpcUtil;
 import io.deephaven.grpc_api_client.barrage.chunk.ChunkInputStreamGenerator;
 import io.deephaven.grpc_api_client.table.BarrageTable;
@@ -28,6 +37,7 @@ import io.deephaven.grpc_api_client.util.BarrageProtoUtil;
 import io.deephaven.grpc_api_client.util.FlatBufferIteratorAdapter;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.Logger;
+import io.grpc.stub.ServerCallStreamObserver;
 import io.grpc.stub.StreamObserver;
 import org.apache.arrow.flatbuf.Message;
 import org.apache.arrow.flatbuf.MessageHeader;
@@ -40,9 +50,14 @@ import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
+import java.util.ArrayDeque;
+import java.util.BitSet;
 import java.util.Iterator;
+import java.util.Queue;
 
 public class ArrowFlightUtil {
+    public static final int DEFAULT_UPDATE_INTERVAL_MS = Configuration.getInstance().getIntegerWithDefault("barrage.updateInterval", 1000);
+
     private static final int TAG_TYPE_BITS = 3;
     private static final BarrageMessage.ModColumnData[] ZERO_MOD_COLUMNS = new BarrageMessage.ModColumnData[0];
 
@@ -120,7 +135,7 @@ public class ArrowFlightUtil {
     /**
      * This is a stateful observer; a DoPut stream begins with its schema.
      */
-    public static class DoPutObserver extends SingletonLivenessManager implements Closeable {
+    public static class DoPutObserver extends SingletonLivenessManager implements BrowserStream.Marshaller<MessageInfo>, StreamObserver<InputStream>, Closeable {
 
         private final SessionState session;
         private final TicketRouter ticketRouter;
@@ -142,9 +157,21 @@ public class ArrowFlightUtil {
             this.observer = observer;
 
             this.session.addOnCloseCallback(this);
+            if (observer instanceof ServerCallStreamObserver) {
+                ((ServerCallStreamObserver<Flight.PutResult>) observer).setOnCancelHandler(this::onCancel);
+            }
         }
 
-        public void process(final MessageInfo mi) {
+        // this is the entry point for client-streams
+        @Override
+        public void onNext(final InputStream request) {
+            GrpcUtil.rpcWrapper(log, observer, () -> {
+                onMessageReceived(parseProtoMessage(request));
+            });
+        }
+
+        // this is the entry point for browser streams; and we delegate to it for client-streams
+        public void onMessageReceived(final MessageInfo mi) {
             if (mi.descriptor != null) {
                 if (resultExportBuilder != null) {
                     throw GrpcUtil.statusRuntimeException(Code.INVALID_ARGUMENT, "Only one descriptor definition allowed");
@@ -251,7 +278,7 @@ public class ArrowFlightUtil {
             }
         }
 
-        public void onComplete() {
+        public void onCompleted() {
             GrpcUtil.rpcWrapper(log, observer, () -> {
                 if (resultExportBuilder == null) {
                     throw GrpcUtil.statusRuntimeException(Code.INVALID_ARGUMENT, "Result flight descriptor never provided");
@@ -291,6 +318,253 @@ public class ArrowFlightUtil {
 
             // retain reference until we can pass this result to be owned by the export object
             resultTable.retainReference();
+        }
+    }
+
+    /**
+     * Helper class that maintains a subscription whether it was created by a bi-directional stream request or the
+     * no-client-streaming request. If the SubscriptionRequest sets the sequence, then it treats sequence as a watermark
+     * and will not send out-of-order requests (due to out-of-band requests). The client should already anticipate
+     * subscription changes may be coalesced by the BarrageMessageProducer.
+     */
+    public static class DoExchangeMarshaller<Options, View> extends SingletonLivenessManager implements BrowserStream.Marshaller<MessageInfo>, StreamObserver<InputStream>, Closeable {
+        @AssistedFactory
+        public interface Factory<Options, View> {
+            DoExchangeMarshaller<Options, View> openExchange(SessionState session, StreamObserver<InputStream> observer);
+        }
+
+        private final String myPrefix;
+        private final SessionState session;
+
+        private long seqWatermark;
+        private boolean isViewport;
+        private BarrageMessageProducer<Options, View> bmp;
+        private Queue<BarrageMessageWrapper> preExportSubscriptions;
+
+        private final StreamObserver<View> listener;
+
+        private boolean isClosed = false;
+        private SessionState.ExportObject<DoExchangeMarshaller<Options, View>> subscriptionExport;
+
+        private final TicketRouter ticketRouter;
+        private final BarrageMessageProducer.Operation.Factory<Options, View> operationFactory;
+        private final BarrageMessageProducer.Adapter<StreamObserver<InputStream>, StreamObserver<View>> listenerAdapter;
+        private final BarrageMessageProducer.Adapter<BarrageSubscriptionRequest, Options> optionsAdapter;
+
+        @AssistedInject
+        public DoExchangeMarshaller(
+                final TicketRouter ticketRouter,
+                final BarrageMessageProducer.Operation.Factory<Options, View> operationFactory,
+                final BarrageMessageProducer.Adapter<StreamObserver<InputStream>, StreamObserver<View>> listenerAdapter,
+                final BarrageMessageProducer.Adapter<BarrageSubscriptionRequest, Options> optionsAdapter,
+                @Assisted final SessionState session, @Assisted final StreamObserver<InputStream> responseObserver) {
+
+            this.myPrefix = "SubscriptionObserver{" + Integer.toHexString(System.identityHashCode(this)) + "}: ";
+            this.ticketRouter = ticketRouter;
+            this.operationFactory = operationFactory;
+            this.listenerAdapter = listenerAdapter;
+            this.optionsAdapter = optionsAdapter;
+            this.session = session;
+            this.listener = listenerAdapter.adapt(responseObserver);
+
+            this.session.addOnCloseCallback(this);
+            if (responseObserver instanceof ServerCallStreamObserver) {
+                ((ServerCallStreamObserver<InputStream>) responseObserver).setOnCancelHandler(this::tryClose);
+            }
+        }
+
+        // this entry is used for client-streaming requests
+        @Override
+        public void onNext(final InputStream request) {
+            GrpcUtil.rpcWrapper(log, listener, () -> {
+                try {
+                    onMessageReceived(parseProtoMessage(request));
+                } catch (final IOException unexpected) {
+                    throw GrpcUtil.securelyWrapError(log, unexpected);
+                }
+            });
+        }
+
+        // this is the entry point for browser streams; and client-streams delegate to this
+        public void onMessageReceived(final MessageInfo message) {
+
+        }
+
+        private synchronized void oldOnNext(final Flight.FlightData subscriptionRequest) {
+            GrpcUtil.rpcWrapper(log, listener, () -> {
+                final BarrageMessageWrapper msg = BarrageMessageWrapper.getRootAsBarrageMessageWrapper(subscriptionRequest.getAppMetadata().asReadOnlyByteBuffer());
+
+                if (bmp == null) {
+                    synchronized (this) {
+                        if (bmp == null) {
+                            queueRequest(msg);
+                            return;
+                        }
+                    }
+                }
+
+                apply(msg);
+            });
+        }
+
+
+        /**
+         * Update the existing subscription to match the new request.
+         * @param msg the requested view change
+         */
+        private void apply(final BarrageMessageWrapper msg) {
+            if (msg.magic() != BarrageStreamGenerator.FLATBUFFER_MAGIC || msg.msgType() != BarrageMessageType.BarrageSubscriptionRequest) {
+                return;
+            }
+
+            if (seqWatermark > 0 && seqWatermark >= msg.sequence()) {
+                return;
+            }
+            seqWatermark = msg.sequence();
+            log.info().append(myPrefix).append("applying subscription request w/seq ").append(seqWatermark).endl();
+
+            final BarrageSubscriptionRequest subscriptionRequest = BarrageSubscriptionRequest.getRootAsBarrageSubscriptionRequest(msg.msgPayloadAsByteBuffer());
+            final boolean hasColumns = subscriptionRequest.columnsVector() != null;
+            final BitSet columns = hasColumns ? BitSet.valueOf(subscriptionRequest.columnsAsByteBuffer()) : new BitSet();
+
+            final boolean hasViewport = subscriptionRequest.viewportVector() != null;
+            final Index viewport = isViewport ? BarrageProtoUtil.toIndex(subscriptionRequest.viewportAsByteBuffer()) : null;
+
+            final boolean subscriptionFound;
+            if (isViewport && hasColumns && hasViewport) {
+                subscriptionFound = bmp.updateViewportAndColumns(listener, viewport, columns);
+            } else if (isViewport && hasViewport) {
+                subscriptionFound = bmp.updateViewport(listener, viewport);
+            } else if (hasColumns) {
+                subscriptionFound = bmp.updateSubscription(listener, columns);
+            } else {
+                subscriptionFound = true;
+            }
+
+            if (!subscriptionFound) {
+                throw GrpcUtil.statusRuntimeException(Code.INTERNAL, "Subscription was not found.");
+            }
+        }
+
+        private synchronized void queueRequest(final BarrageMessageWrapper msg) {
+            if (preExportSubscriptions != null) {
+                preExportSubscriptions.add(msg);
+                return;
+            }
+
+            preExportSubscriptions = new ArrayDeque<>();
+            preExportSubscriptions.add(msg);
+
+            final BarrageSubscriptionRequest subscriptionRequest = BarrageSubscriptionRequest.getRootAsBarrageSubscriptionRequest(msg.msgPayloadAsByteBuffer());
+            if (subscriptionRequest.ticketVector() == null) {
+                listener.onError(GrpcUtil.statusRuntimeException(Code.INVALID_ARGUMENT, "Ticket not specified."));
+                return;
+            }
+
+            final SessionState.ExportObject<Object> parent = ticketRouter.resolve(session, subscriptionRequest.ticketAsByteBuffer());
+
+            final SessionState.ExportBuilder<DoExchangeMarshaller<Options, View>> exportBuilder;
+            if (msg.rpcTicketVector() != null) {
+                exportBuilder = session.newExport(ExportTicketHelper.ticketToExportId(msg.rpcTicketAsByteBuffer()));
+            } else {
+                exportBuilder = session.nonExport();
+            }
+
+            subscriptionExport = exportBuilder
+                    .require(parent)
+                    .onError(listener::onError)
+                    .submit(() -> {
+                        synchronized (DoExchangeMarshaller.this) {
+                            subscriptionExport = null;
+
+                            if (isClosed) {
+                                return null;
+                            }
+
+                            final Object export = parent.get();
+                            if (export instanceof QueryTable) {
+                                final QueryTable table = (QueryTable) export;
+                                long updateIntervalMs = subscriptionRequest.updateIntervalMs();
+                                if (updateIntervalMs == 0) {
+                                    updateIntervalMs = DEFAULT_UPDATE_INTERVAL_MS;
+                                }
+                                bmp = table.getResult(operationFactory.create(table, updateIntervalMs));
+                                manage(bmp);
+                            } else {
+                                listener.onError(GrpcUtil.statusRuntimeException(Code.FAILED_PRECONDITION, "Ticket ("
+                                        + ExportTicketHelper.toReadableString(subscriptionRequest.ticketAsByteBuffer())
+                                        + ") is not a subscribable table."));
+                                return null;
+                            }
+
+                            log.info().append(myPrefix).append("processing initial subscription").endl();
+                            final BarrageMessageWrapper initial = preExportSubscriptions.remove();
+                            seqWatermark = initial.sequence();
+
+                            final boolean hasColumns = subscriptionRequest.columnsVector() != null;
+                            final BitSet columns = hasColumns ? BitSet.valueOf(subscriptionRequest.columnsAsByteBuffer()) : new BitSet();
+
+                            isViewport = subscriptionRequest.viewportVector() != null;
+                            final Index viewport = isViewport ? BarrageProtoUtil.toIndex(subscriptionRequest.viewportAsByteBuffer()) : null;
+
+                            if (!bmp.addSubscription(listener, optionsAdapter.adapt(subscriptionRequest), columns, viewport)) {
+                                throw new IllegalStateException("listener is already a subscriber!");
+                            }
+
+                            for (final BarrageMessageWrapper request : preExportSubscriptions) {
+                                apply(request);
+                            }
+
+                            // we will now process requests as they are received
+                            preExportSubscriptions = null;
+                            return DoExchangeMarshaller.this;
+                        }
+                    });
+        }
+
+        public void onCancel() {
+            log.debug().append(myPrefix).append("cancel requested").endl();
+            tryClose();
+        }
+
+        @Override
+        public void onError(final Throwable t) {
+            log.error().append(myPrefix).append("unexpected error; force closing subscription: caused by ").append(t).endl();
+            tryClose();
+        }
+
+        @Override
+        public void onCompleted() {
+            log.error().append(myPrefix).append("client stream closed subscription").endl();
+            tryClose();
+        }
+
+        @Override
+        public void close() {
+            synchronized (this) {
+                if (isClosed) {
+                    return;
+                }
+
+                isClosed = true;
+            }
+
+            if (subscriptionExport != null) {
+                subscriptionExport.cancel();
+                subscriptionExport = null;
+            }
+
+            if (bmp != null) {
+                bmp.removeSubscription(listener);
+                bmp = null;
+            }
+            release();
+        }
+
+        private void tryClose() {
+            if (session.removeOnCloseCallback(this) != null) {
+                close();
+            }
         }
     }
 }
